+++ kommons-uri/src/commonMain/kotlin/com/bkahlert/kommons/uri/Strings.kt
@@ -1,8 +1,5 @@
 package com.bkahlert.kommons.uri
 
-import io.ktor.utils.io.charsets.Charset
-import io.ktor.utils.io.charsets.encodeToByteArray
-
 /**
  * Returns the character sequence parsed as a [Uri]
  * as specified in [RFC3986 Appendix B](https://www.rfc-editor.org/rfc/rfc3986#appendix-B), or
@@ -16,9 +13,3 @@ public fun CharSequence.toUri(): Uri = Uri.parse(this)
  * `null` otherwise.
  */
 public fun CharSequence.toUriOrNull(): Uri? = Uri.parseOrNull(this)
-
-/**
- * Encodes this character sequence using the specified [charset].
- */
-internal fun CharSequence.encodeToByteArray(charset: Charset): ByteArray =
-    charset.newEncoder().encodeToByteArray(this)