+++ kommons-uri/src/commonTest/kotlin/com/bkahlert/kommons/uri/UrisKtTest.kt
@@ -47,6 +47,87 @@ class UrisKtTest {
         EMPTY_URI.fragmentParameters shouldBe Parameters.Empty
     }
 
+    /**
+     * Example base [Uri] as described in [RFC3986 section 5.4](https://www.rfc-editor.org/rfc/rfc3986#section-5.4)
+     */
+    private val baseUri = Uri.parse("http://a/b/c/d;p?q")
+
+    /**
+     * References resolution examples as described in [RFC3986 section 5.4.1](https://www.rfc-editor.org/rfc/rfc3986#section-5.4.1)
+     */
+    @Test
+    fun resolve_normal_examples() = testAll(
+        "g:h" to "g:h",
+        "g" to "http://a/b/c/g",
+        "./g" to "http://a/b/c/g",
+        "g/" to "http://a/b/c/g/",
+        "/g" to "http://a/g",
+        "//g" to "http://g",
+        "?y" to "http://a/b/c/d;p?y",
+        "g?y" to "http://a/b/c/g?y",
+        "#s" to "http://a/b/c/d;p?q#s",
+        "g#s" to "http://a/b/c/g#s",
+        "g?y#s" to "http://a/b/c/g?y#s",
+        ";x" to "http://a/b/c/;x",
+        "g;x" to "http://a/b/c/g;x",
+        "g;x?y#s" to "http://a/b/c/g;x?y#s",
+        "" to "http://a/b/c/d;p?q",
+        "." to "http://a/b/c/",
+        "./" to "http://a/b/c/",
+        ".." to "http://a/b/",
+        "../" to "http://a/b/",
+        "../g" to "http://a/b/g",
+        "../.." to "http://a/",
+        "../../" to "http://a/",
+        "../../g" to "http://a/g",
+    ) { (relativeReference, targetUriString) ->
+        val targetUri = Uri.parse(targetUriString)
+        val uriReference = Uri.parse(relativeReference)
+
+        baseUri.resolve(relativeReference) shouldBe targetUri
+        baseUri.resolve(uriReference) shouldBe targetUri
+        uriReference.resolveTo(baseUri) shouldBe targetUri
+    }
+
+    /**
+     * References resolution examples as described in [RFC3986 section 5.4.2](https://www.rfc-editor.org/rfc/rfc3986#section-5.4.2)
+     */
+    @Test
+    fun resolve_abnormal_examples() = testAll(
+        "../../../g" to "http://a/g",
+        "../../../../g" to "http://a/g",
+
+        "/./g" to "http://a/g",
+        "/../g" to "http://a/g",
+        "g." to "http://a/b/c/g.",
+        ".g" to "http://a/b/c/.g",
+        "g.." to "http://a/b/c/g..",
+        "..g" to "http://a/b/c/..g",
+
+        "./../g" to "http://a/b/g",
+        "./g/." to "http://a/b/c/g/",
+        "g/./h" to "http://a/b/c/g/h",
+        "g/../h" to "http://a/b/c/h",
+        "g;x=1/./y" to "http://a/b/c/g;x=1/y",
+        "g;x=1/../y" to "http://a/b/c/y",
+
+        "g?y/./x" to "http://a/b/c/g?y/./x",
+        "g?y/../x" to "http://a/b/c/g?y/../x",
+        "g#s/./x" to "http://a/b/c/g#s/./x",
+        "g#s/../x" to "http://a/b/c/g#s/../x",
+    ) { (relativeReference, targetUriString) ->
+        baseUri.resolve(relativeReference) shouldBe Uri.parse(targetUriString)
+    }
+
+    /**
+     * References resolution examples as described in [RFC3986 section 5.4.2](https://www.rfc-editor.org/rfc/rfc3986#section-5.4.2)
+     */
+    @Test
+    fun resolve_reference_with_scheme() = testAll {
+        Uri.parse("http:g").resolveTo(baseUri, strict = true) shouldBe Uri.parse("http:g")
+        Uri.parse("http:g").resolveTo(baseUri, strict = false) shouldBe Uri.parse("http://a/b/c/g")
+    }
+
 
     @Test
     fun div() = testAll {